+++ config.json
@@ -1,0 +1,8 @@
+{ 
+    "tokens"   : {
+        "mekhane": "YOUR-TOKEN-HERE"
+},
+    "prefixes" : {
+        "mekhane": "&"
+  }
+}
