--- files/data.json
@@ -1,6 +1,0 @@
-{ 
-    "mekhane" : {
-        "scriptbans": [""],
-        "blacklist" : [""] 
-    }
-}
